requirements.txt
flask-cors
flask
flask-login
flask-migrate
flask-sqlalchemy
flask-wtf
passlib
psycopg2